feat: reorganize backend structure
