+++ .gitignore
@@ -14,6 +14,7 @@ dist-ssr
 
 # Backend
 backend/venv
+*__pycache__
 
 # Editor directories and files
 .vscode/*
